+++ core/app-version.js
@@ -1,1 +1,1 @@
-export const APP_VERSION = '20260629-5';
+export const APP_VERSION = '20260630-1';